test - move test about addRate
